[ELY-40] Split out the Base64 handling into it's own utility class.
